backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-socketio==5.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.0.3
openai==1.3.5
stripe==7.4.0
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
celery==5.3.4
flower==2.0.1
sentry-sdk==1.38.0
prometheus-client==0.19.0
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
psutil==5.9.6